codex-cli/pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "codex-cli"
version = "0.1.0"
description = "A powerful CLI for interacting with multiple LLM providers"
readme = "README.md"
license = {text = "MIT"}
authors = [
    {name = "Codex CLI Team"}
]
keywords = ["cli", "llm", "ai", "chatgpt", "gemini", "groq", "mistral", "anthropic"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
]
dependencies = [
    "requests>=2.28.0",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
google = ["google-generativeai>=0.3.0"]
cohere = ["cohere>=5.0.0"]
all = [
    "google-generativeai>=0.3.0",
    "cohere>=5.0.0",
]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
]

[project.scripts]
codex = "codex_cli.cli:main"

[project.urls]
Homepage = "https://github.com/yourusername/codex-cli"
Documentation = "https://github.com/yourusername/codex-cli#readme"
Repository = "https://github.com/yourusername/codex-cli.git"
Issues = "https://github.com/yourusername/codex-cli/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.black]
line-length = 100
target-version = ['py38', 'py39', 'py310', 'py311', 'py312']

[tool.ruff]
line-length = 100
select = ["E", "F", "W", "I"]
